move constants to their own directory
