+++ src/Opdex.Platform.Application/EntryHandlers/Blocks/ProcessLatestBlocksCommandHandler.cs
@@ -15,6 +15,7 @@ using Opdex.Platform.Application.Abstractions.Queries.Markets;
 using Opdex.Platform.Application.Abstractions.Queries.Tokens;
 using Opdex.Platform.Application.Abstractions.Queries.Tokens.Snapshots;
 using Opdex.Platform.Common;
+using Opdex.Platform.Common.Constants;
 using Opdex.Platform.Common.Enums;
 
 namespace Opdex.Platform.Application.EntryHandlers.Blocks
