+++ src/Opdex.Platform.Application/EntryHandlers/Transactions/Wallet/CreateWalletStopStakingTransactionCommandHandler.cs
@@ -5,6 +5,7 @@ using MediatR;
 using Opdex.Platform.Application.Abstractions.Commands.Transactions.Wallet;
 using Opdex.Platform.Application.Abstractions.EntryCommands.Transactions.Wallet;
 using Opdex.Platform.Common;
+using Opdex.Platform.Common.Constants;
 using Opdex.Platform.Common.Extensions;
 
 namespace Opdex.Platform.Application.EntryHandlers.Transactions.Wallet
