+++ src/Opdex.Platform.Application/EntryHandlers/Vaults/CreateWalletCreateVaultCertificateCommandHandler.cs
@@ -2,6 +2,7 @@ using MediatR;
 using Opdex.Platform.Application.Abstractions.Commands.Vaults;
 using Opdex.Platform.Application.Abstractions.EntryCommands.Vaults;
 using Opdex.Platform.Common;
+using Opdex.Platform.Common.Constants;
 using Opdex.Platform.Common.Extensions;
 using System.Threading;
 using System.Threading.Tasks;
