+++ src/Opdex.Platform.Common/Extensions/StringExtensions.cs
@@ -1,3 +1,4 @@
+using Opdex.Platform.Common.Constants;
 using System;
 using System.Text.RegularExpressions;
