+++ src/Opdex.Platform.Domain/Models/Tokens/TokenSnapshot.cs
@@ -1,5 +1,6 @@
 using System;
 using Opdex.Platform.Common;
+using Opdex.Platform.Common.Constants;
 using Opdex.Platform.Common.Enums;
 using Opdex.Platform.Common.Extensions;
 using Opdex.Platform.Domain.Models.OHLC;
