+++ src/Opdex.Platform.Infrastructure/Clients/CirrusFullNodeApi/Handlers/Tokens/CallCirrusGetSrcTokenDetailsByAddressQueryHandler.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Opdex.Platform.Common;
+using Opdex.Platform.Common.Constants;
 using Opdex.Platform.Common.Extensions;
 using Opdex.Platform.Domain.Models.Tokens;
 using Opdex.Platform.Infrastructure.Abstractions.Clients.CirrusFullNodeApi.Models;
