+++ test/Opdex.Platform.Application.Tests/EntryHandlers/Vaults/CreateWalletCreateVaultCertificateCommandHandlerTests.cs
@@ -5,6 +5,7 @@ using Opdex.Platform.Application.Abstractions.Commands.Vaults;
 using Opdex.Platform.Application.Abstractions.EntryCommands.Vaults;
 using Opdex.Platform.Application.EntryHandlers.Vaults;
 using Opdex.Platform.Common;
+using Opdex.Platform.Common.Constants;
 using Opdex.Platform.Common.Extensions;
 using System.Threading;
 using System.Threading.Tasks;
