+++ test/Opdex.Platform.Common.Tests/StringExtensionsTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Numerics;
 using FluentAssertions;
+using Opdex.Platform.Common.Constants;
 using Opdex.Platform.Common.Extensions;
 using Xunit;
 using Xunit.Abstractions;
