+++ test/Opdex.Platform.Domain.Tests/Models/Pools/Snapshots/LiquidityPoolSnapshotTests.cs
@@ -1,6 +1,7 @@
 using System;
 using FluentAssertions;
 using Opdex.Platform.Common;
+using Opdex.Platform.Common.Constants;
 using Opdex.Platform.Common.Enums;
 using Opdex.Platform.Common.Extensions;
 using Opdex.Platform.Domain.Models.OHLC;
